Move RequestLoggerMiddleware to internal package
